Move sleep timer & settings button to details menu
