+++ lib/screens/channel/channel.dart
@@ -16,12 +16,9 @@ import 'package:frosty/screens/channel/chat/stores/chat_store.dart';
 import 'package:frosty/screens/channel/video/video.dart';
 import 'package:frosty/screens/channel/video/video_overlay.dart';
 import 'package:frosty/screens/channel/video/video_store.dart';
-import 'package:frosty/screens/settings/settings.dart';
 import 'package:frosty/screens/settings/stores/auth_store.dart';
 import 'package:frosty/screens/settings/stores/settings_store.dart';
 import 'package:frosty/widgets/app_bar.dart';
-import 'package:frosty/widgets/bottom_sheet.dart';
-import 'package:heroicons/heroicons.dart';
 import 'package:provider/provider.dart';
 import 'package:simple_pip_mode/actions/pip_actions_layout.dart';
 import 'package:simple_pip_mode/pip_widget.dart';
@@ -76,31 +73,12 @@ class _VideoChatState extends State<VideoChat> {
   Widget build(BuildContext context) {
     final settingsStore = _chatStore.settings;
 
-    void showSettings() => showModalBottomSheet(
-          backgroundColor: Colors.transparent,
-          isScrollControlled: true,
-          context: context,
-          builder: (context) => FrostyBottomSheet(
-            child: SizedBox(
-              height: MediaQuery.of(context).size.height * 0.8,
-              child: Settings(settingsStore: settingsStore),
-            ),
-          ),
-        );
-
     final appBar = FrostyAppBar(
       title: Text(
         regexEnglish.hasMatch(_chatStore.displayName)
             ? _chatStore.displayName
             : '${_chatStore.displayName} (${_chatStore.channelName})',
       ),
-      actions: [
-        IconButton(
-          tooltip: 'Settings',
-          icon: const HeroIcon(HeroIcons.cog6Tooth, style: HeroIconStyle.solid),
-          onPressed: showSettings,
-        ),
-      ],
     );
 
     final player = GestureDetector(
@@ -113,7 +91,6 @@ class _VideoChatState extends State<VideoChat> {
 
     final videoOverlay = VideoOverlay(
       videoStore: _videoStore,
-      onSettingsPressed: showSettings,
     );
 
     final overlay = GestureDetector(
